+++ sambaplayersdk/src/main/java/com/sambatech/player/offline/model/SambaDownloadListener.java
@@ -1,0 +1,5 @@
+package com.sambatech.player.offline.model;
+
+public interface SambaDownloadListener {
+    void onDownloadStateChanged(DownloadState downloadState);
+}
